cleanup

+++ crates/types/src/camera_matrix.rs
@@ -117,7 +117,7 @@ mod tests {
 
         let focals_scaled = image_size_abs
             .inner
-            .zip_map(&focals, |dim, focal| dim * focal);
+            .zip_map(&focals, |dim, focal| dim as f32 * focal);
 
         assert_relative_eq!(
             old_fov(focals),
